order_manage/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.2
django-extensions==3.1.3
djangorestframework==3.12.4
Pillow==8.2.0
pkg-resources==0.0.0
pycodestyle==2.7.0
pytz==2021.1
reportlab==3.5.67
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
